requirements.txt
instagrapi>=2.1.0
requests>=2.31.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
